[PATCH] netfilter: Fix build failure when ipv6 but xt_tproxy is built in
net/built-in.o: In function `tproxy_tg6_v1':
/home/rostedt/work/autotest/nobackup/linux-test.git/net/netfilter/xt_TPROXY.c:288: undefined reference to `ipv6_find_hdr'

This happened because the xt_TPROXY code was compiled into the kernel
proper, but the ipv6 netfilter was compiled as a module. The fix is to
only enter the code that calls ipv6_find_hdr if ipv6 netfilter is
compiled into the kernel, or if it is a module, so is the xt_TPROXY
code.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>